refactor: extract BaseCacheFile for reuse
Extract common file caching functionality from ConfigCacheFile into
a new BaseCacheFile base class. This enables reuse for the upcoming
precomputed configuration cache without code duplication.

- Add BaseCacheFile with common read/write/delete operations
- Refactor ConfigCacheFile to extend BaseCacheFile
- No functional changes to existing behavior
