+++ eppo/src/main/java/cloud/eppo/android/ConfigCacheFile.java
@@ -1,70 +1,15 @@
 package cloud.eppo.android;
 
 import android.app.Application;
-import java.io.BufferedReader;
-import java.io.BufferedWriter;
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileNotFoundException;
-import java.io.FileOutputStream;
-import java.io.FileReader;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStream;
 
-public class ConfigCacheFile {
-  private final File cacheFile;
+/** Disk cache file for standard flag configuration. */
+public class ConfigCacheFile extends BaseCacheFile {
 
   public ConfigCacheFile(Application application, String fileNameSuffix) {
-    File filesDir = application.getFilesDir();
-    cacheFile = new File(filesDir, cacheFileName(fileNameSuffix));
+    super(application, cacheFileName(fileNameSuffix));
   }
 
   public static String cacheFileName(String suffix) {
     return "eppo-sdk-config-v4-flags-" + suffix + ".json";
-  }
-
-  public boolean exists() {
-    return cacheFile.exists();
-  }
-
-  /**
-   * @noinspection ResultOfMethodCallIgnored
-   */
-  public void delete() {
-    if (cacheFile.exists()) {
-      cacheFile.delete();
-    }
-  }
-
-  /** Useful for passing in as a writer for gson serialization */
-  public BufferedWriter getWriter() throws IOException {
-    return new BufferedWriter(new FileWriter(cacheFile));
-  }
-
-  public OutputStream getOutputStream() throws FileNotFoundException {
-    return new FileOutputStream(cacheFile);
-  }
-
-  public InputStream getInputStream() throws FileNotFoundException {
-    return new FileInputStream(cacheFile);
-  }
-
-  /** Useful for passing in as a reader for gson deserialization */
-  public BufferedReader getReader() throws IOException {
-    return new BufferedReader(new FileReader(cacheFile));
-  }
-
-  /** Useful for mocking caches in automated tests */
-  public void setContents(String contents) {
-    delete();
-    try {
-      BufferedWriter writer = getWriter();
-      writer.write(contents);
-      writer.close();
-    } catch (IOException ex) {
-      throw new RuntimeException(ex);
-    }
   }
 }
